Correccion de error en response de dates

+++ src/controllers/drivers_in_teams.controller.js
@@ -8,7 +8,7 @@ const getDriversInTeamByTeamId = async (req, res) => {
         if (response.rows.length > 0 ){
             res.status(200).json(response.rows[0]);
         }else {
-            res.status(405).json({error: 'not found.'});
+            res.status(404).json({error: 'not found.'});
         }
     }
 }
@@ -30,12 +30,11 @@ const createDriversInTeam = async (req, res) => {
         }else{
             res.status(404).json({
                 error: 'invalid parameters',
-                description: 'One of the parameters does not exists.', 
             });
         }
 
     }catch(err){
-        res.status(405).json({
+        res.status(400).json({
             error: 'creation failed',
         }); 
     }
@@ -46,14 +45,14 @@ const updateDriversInTeam = async (req, res) => {
         if (isNaN([req.params.id])){
             res.status(404).json({
                 error: 'invalid parameter',
-                description: 'The team does not exists.'
             }); 
         }else{
+            const checkExistsTeam = await db.query('SELECT * FROM drivers_in_team WHERE id = $1', [req.params.id]);
             const {driver_1_id, driver_2_id} = req.body;
             const checkExistingDriver1 = await db.query('SELECT * FROM drivers WHERE id = $1',[driver_1_id]);
             const checkExistingDriver2 = await db.query('SELECT * FROM drivers WHERE id = $1',[driver_2_id]);
             
-            if (checkExistingDriver1.rowCount>0 && checkExistingDriver2.rowCount>0){
+            if (checkExistsTeam.rowCount>0 && checkExistingDriver1.rowCount>0 && checkExistingDriver2.rowCount>0){
                 const response = await db.query('UPDATE drivers_in_teams SET driver_1_id = $1, driver_2_id = $2 WHERE id = $3', [
                     driver_1_id, 
                     driver_2_id, 
@@ -63,14 +62,12 @@ const updateDriversInTeam = async (req, res) => {
             }else{
                 res.status(404).json({
                     error: 'invalid parameters',
-                    description: 'One of the parameters does not exists.', 
                 });
             }
         }
     }catch(err){
-        res.status(405).json({
+        res.status(400).json({
             error: 'update failed',
-            description: err.message, 
         }); 
     }
 }
@@ -80,13 +77,17 @@ const deleteDriversInTeam = async (req, res) => {
         if (isNaN([req.params.id])){
             res.status(404).json({error: 'parameter not valid.'}); 
         }else{
-            const response = await db.query('DELETE FROM drivers_in_teams WHERE id = $1',[req.params.id]);
-            res.status(200).json({success: 'true'});
+            const checkExistsTeam = await db.query('SELECT * FROM drivers_in_team WHERE id = $1', [req.params.id]);
+            if (checkExistsTeam.rowCount>0){
+                const response = await db.query('DELETE FROM drivers_in_teams WHERE id = $1',[req.params.id]);
+                res.status(200).json({success: 'true'});
+            }else{
+                res.status(404).json({error: 'parameter not valid.'}); 
+            }
         }
     }catch(err){
-        res.status(405).json({
+        res.status(400).json({
             error: 'delete failed',
-            description: err.message, 
         }); 
     }
 }
